+++ src/linux_server_bot/shared/actions/sysinfo.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 
 import logging
-import time
 
+from linux_server_bot.shared.cpu import read_cpu_percent
 from linux_server_bot.shared.shell import run_command, run_shell
 
 logger = logging.getLogger(__name__)
@@ -118,43 +118,12 @@ def get_sysinfo_text() -> str:
     return "\n".join(out)
 
 
-def _read_cpu_times() -> tuple[int, int] | None:
-    """Read aggregate CPU jiffies from /proc/stat. Returns (total, idle) or None."""
-    result = run_shell("head -n 1 /proc/stat")
-    if not result.success:
-        return None
-    parts = result.stdout.split()
-    if len(parts) < 5 or parts[0] != "cpu":
-        return None
-    try:
-        fields = [int(x) for x in parts[1:]]
-    except ValueError:
-        return None
-    idle = fields[3]
-    total = sum(fields)
-    return total, idle
-
-
 def get_cpu_usage() -> dict:
-    """Get CPU usage percentage from /proc/stat over a 1-second window.
-
-    Replaces the previous ``top -bn 1 | awk`` parsing which mis-read idle as 0
-    whenever any %Cpu(s) field reached 100.0 (top's fixed-width column shifts
-    by one when the leading space disappears, so awk's $8 became "ni," instead
-    of the idle value).
-    """
-    first = _read_cpu_times()
-    if first is None:
+    """Get CPU usage percentage from /proc/stat over a 1-second window."""
+    pct = read_cpu_percent()
+    if pct is None:
         return {"cpu_percent": None, "success": False, "error": "Could not read /proc/stat"}
-    time.sleep(1)
-    second = _read_cpu_times()
-    if second is None:
-        return {"cpu_percent": None, "success": False, "error": "Could not read /proc/stat"}
-    total_delta = second[0] - first[0]
-    idle_delta = second[1] - first[1]
-    if total_delta <= 0:
-        return {"cpu_percent": None, "success": False, "error": "Invalid /proc/stat delta"}
-    return {"cpu_percent": round(100.0 * (1.0 - idle_delta / total_delta), 1), "success": True}
+    return {"cpu_percent": pct, "success": True}
 
 
 def get_memory_usage() -> dict:
